update the share token default to be an empty string, and fix a flaky test

+++ MaveSDK/Controllers/MAVEInvitePageViewController.h
@@ -36,6 +36,6 @@
 - (void)showErrorAndResetAfterSendInvitesFailure:(NSError *)error;
 
 // Methods to dismiss self after user done sending invites or user hit cancel
-- (void)dismissSelf:(unsigned int)numberOfInvitesSent;
+- (void)dismissSelf:(NSUInteger)numberOfInvitesSent;
 - (void)dismissAfterCancel;
 @end
